pyproject.toml
[tool.poetry]
name = "gemini-service"
version = "0.1.0"
description = "A service to generate structured product JSON from raw text using the Gemini API."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
google-generativeai = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"